+++ apps/icons-plugin/src/app/components/icon-search/icon-search.component.ts
@@ -2,17 +2,17 @@ import { Component, output, signal } from '@angular/core';
 import { icons } from 'feather-icons';
 
 @Component({
-    selector: 'app-icon-search',
-    imports: [],
-    styleUrl: './icon-search.component.css',
-    template: `
+  selector: 'app-icon-search',
+  imports: [],
+  styleUrl: './icon-search.component.css',
+  template: `
     <input
       class="search-icon"
       type="search"
       placeholder="Search an icon"
       (input)="onSearchIcons($event)"
     />
-  `
+  `,
 })
 export class IconSearchComponent {
   public searchIcons = output<string>();
@@ -20,6 +20,6 @@ export class IconSearchComponent {
 
   public onSearchIcons(event: Event): void {
     const target = event.target as HTMLInputElement;
-    this.searchIcons.emit(target?.value.toLowerCase()  || '');
+    this.searchIcons.emit(target?.value.toLowerCase() || '');
   }
 }