+++ libs/plugins-runtime/src/lib/api/openUI.api.ts
@@ -9,7 +9,7 @@ export const openUIApi = z
     z.string(),
     z.enum(['dark', 'light']),
     openUISchema.optional(),
-    z.boolean().optional()
+    z.boolean().optional(),
   )
   .implement((title, url, theme, options, allowDownloads) => {
     return createModal(title, url, theme, options, allowDownloads);